[main] Implement base usage (no flag, no file)
By default, `tee` copies stdin to stdout if no files are given.
Therefore, the initial step is implemented in the same way.

Also, a placeholder signal notification is implemented for SIGINT.
The implementation is the same as how Go handles SIGINT by default.
Based on needs, the implementation may be removed in future.