[flags] Implement -a
With this implementation, the initial stage of the program is finished.

The writes are done in a concurrent fashion for each argument passed to
`gtee`.

`context.Context` is used to cancel the execution when the SIGINT signal
is sent to `gtee`.

The main process is divided into functions to make `gtee` a bit
more readable.

Possible errors coming from syscalls are wrapped with their context
(reads, writes) to help with diagnosing.