[tests] Add tests for exit codes
POSIX specification says that `tee` should exit with code >0 if any
error happens during execution.
If not, the exit status code should be 0.

Therefore, errChan is redesigned to be used for all errors, not just
stdin.

Also, the main execution point is updated to return a status code, which
is used by os.Exit() in main().

Test cases are added for 0 and >0 code scenarios. Error is triggered by
trying to open a dir as a file.