[tests] Add benchmarks to compare gtee with tee
A simple benchmark test for 2 binaries (`tee`, and `gtee`) are
implemented.

The benchmark first creates a 100MB file via an external command `dd`,
and then sets `stdin` to this file so that both binaries read from the
exact same input.

The binaries write the output to one single file, and stdout is set to
nil to work on a single file only.